:recycle: Move `DataSet` tests to `test_dataset.py`
